+++ crates/wdl-engine/src/backend/apptainer.rs
@@ -29,7 +29,7 @@ use tracing::warn;
 
 use crate::Value;
 use crate::backend::ExecuteTaskRequest;
-use crate::config::Config;
+use crate::config::ApptainerConfig;
 use crate::config::DEFAULT_TASK_SHELL;
 use crate::v1::requirements::ContainerSource;
 
@@ -47,6 +47,12 @@ const GUEST_STDOUT_PATH: &str = "/mnt/task/stdout";
 
 /// The path to the container's stderr.
 const GUEST_STDERR_PATH: &str = "/mnt/task/stderr";
+
+/// The environment variable prefix for Apptainer.
+const APPTAINER_ENV_PREFIX: &str = "APPTAINERENV";
+
+/// The environment variable prefix for Singularity.
+const SINGULARITY_ENV_PREFIX: &str = "SINGULARITYENV";
 
 /// Represents the Apptainer container runtime.
 #[derive(Debug)]
@@ -90,13 +96,14 @@ impl ApptainerRuntime {
     /// GPFS.
     pub async fn generate_script(
         &self,
-        config: &Config,
+        config: &ApptainerConfig,
+        shell: Option<&str>,
         request: &ExecuteTaskRequest<'_>,
-        extra_args: impl Iterator<Item = &str>,
         token: CancellationToken,
     ) -> Result<Option<String>> {
         let path = match self
             .pull_image(
+                &config.executable,
                 request
                     .constraints
                     .container
@@ -111,7 +118,7 @@ impl ApptainerRuntime {
         };
 
         Ok(Some(
-            self.generate_apptainer_script(config, &path, request, extra_args)
+            self.generate_apptainer_script(config, shell, &path, request)
                 .await?,
         ))
     }
@@ -123,10 +130,10 @@ impl ApptainerRuntime {
     /// be called from outside this module.
     async fn generate_apptainer_script(
         &self,
-        config: &Config,
+        config: &ApptainerConfig,
+        shell: Option<&str>,
         container_sif: &Path,
         request: &ExecuteTaskRequest<'_>,
-        extra_args: impl Iterator<Item = &str>,
     ) -> Result<String> {
         // Create a temp dir for the container's execution within the attempt dir
         // hierarchy. On many HPC systems, `/tmp` is mapped to a relatively
@@ -158,12 +165,18 @@ impl ApptainerRuntime {
                 )
             })?;
 
+        let env_prefix = if config.executable.contains("singularity") {
+            SINGULARITY_ENV_PREFIX
+        } else {
+            APPTAINER_ENV_PREFIX
+        };
+
         let mut apptainer_command = String::new();
         writeln!(&mut apptainer_command, "#!/usr/bin/env bash")?;
         for (k, v) in request.env.iter() {
-            writeln!(&mut apptainer_command, "export APPTAINERENV_{k}={v:?}")?;
+            writeln!(&mut apptainer_command, "export {env_prefix}_{k}={v:?}")?;
         }
-        writeln!(&mut apptainer_command, "apptainer -v exec \\")?;
+        writeln!(&mut apptainer_command, "{} -v exec \\", config.executable)?;
         writeln!(&mut apptainer_command, "--pwd \"{GUEST_WORK_DIR}\" \\")?;
         writeln!(&mut apptainer_command, "--containall --cleanenv \\")?;
         for input in request.backend_inputs {
@@ -218,7 +231,11 @@ impl ApptainerRuntime {
             writeln!(&mut apptainer_command, "--nv \\")?;
         }
 
-        for arg in extra_args {
+        for arg in config
+            .extra_apptainer_exec_args
+            .as_deref()
+            .unwrap_or_default()
+        {
             writeln!(&mut apptainer_command, "{arg} \\")?;
         }
 
@@ -227,7 +244,7 @@ impl ApptainerRuntime {
             &mut apptainer_command,
             "{shell} -c \"\\\"{GUEST_COMMAND_PATH}\\\" > \\\"{GUEST_STDOUT_PATH}\\\" 2> \
              \\\"{GUEST_STDERR_PATH}\\\"\" \\",
-            shell = config.task.shell.as_deref().unwrap_or(DEFAULT_TASK_SHELL)
+            shell = shell.unwrap_or(DEFAULT_TASK_SHELL)
         )?;
         let attempt_dir = request.attempt_dir;
         let apptainer_stdout_path = attempt_dir.join("apptainer.stdout");
@@ -251,6 +268,7 @@ impl ApptainerRuntime {
     /// to the previous location is returned.
     pub(crate) async fn pull_image(
         &self,
+        executable: &str,
         container: &ContainerSource,
         token: CancellationToken,
     ) -> Result<Option<PathBuf>> {
@@ -294,6 +312,7 @@ impl ApptainerRuntime {
             }
 
             let container = format!("{container:#}");
+            let executable = executable.to_string();
 
             Retry::spawn_notify(
                 // TODO ACF 2025-09-22: configure the retry behavior based on actual experience
@@ -304,9 +323,12 @@ impl ApptainerRuntime {
                 ExponentialBackoff::from_millis(50)
                     .max_delay_millis(60_000)
                     .take(10),
-                || Self::try_pull_image(&container, &path),
-                |e: &anyhow::Error, _| {
-                    warn!(e = %e, "`apptainer pull` failed");
+                || Self::try_pull_image(&executable, &container, &path),
+                {
+                    let executable = executable.clone();
+                    move |e: &anyhow::Error, _| {
+                        warn!(e = %e, "`{executable} pull` failed");
+                    }
                 },
             )
             .await
@@ -334,10 +356,14 @@ impl ApptainerRuntime {
     /// whether a failure is transient, but as we gain experience recognizing
     /// its output patterns, we can enhance the fidelity of the error
     /// handling.
-    async fn try_pull_image(image: &str, path: &Path) -> Result<(), RetryError<anyhow::Error>> {
-        info!("spawning `apptainer` to pull image `{image}`");
+    async fn try_pull_image(
+        executable: &str,
+        image: &str,
+        path: &Path,
+    ) -> Result<(), RetryError<anyhow::Error>> {
+        info!("spawning `{executable}` to pull image `{image}`");
 
-        let child = Command::new("apptainer")
+        let child = Command::new(executable)
             .stdin(Stdio::null())
             .stdout(Stdio::piped())
             .stderr(Stdio::piped())
@@ -347,7 +373,7 @@ impl ApptainerRuntime {
             .spawn()
             .with_context(|| {
                 format!(
-                    "failed to spawn `apptainer pull '{path}' '{image}'",
+                    "failed to spawn `{executable} pull '{path}' '{image}'`",
                     path = path.display()
                 )
             })
@@ -357,7 +383,7 @@ impl ApptainerRuntime {
         let output = child
             .wait_with_output()
             .await
-            .context("failed to wait for `apptainer`")
+            .context(format!("failed to wait for `{executable}`"))
             .map_err(RetryError::permanent)?;
         if !output.status.success() {
             let permanent = if let Ok(stderr) = str::from_utf8(&output.stderr) {
@@ -379,7 +405,7 @@ impl ApptainerRuntime {
             };
 
             let e = anyhow!(
-                "`apptainer` failed: {status}: {stderr}",
+                "`{executable}` failed: {status}: {stderr}",
                 status = output.status,
                 stderr = str::from_utf8(&output.stderr)
                     .unwrap_or("<output not UTF-8>")
@@ -398,8 +424,6 @@ impl ApptainerRuntime {
 
 #[cfg(test)]
 mod tests {
-    use std::iter::empty;
-
     use indexmap::IndexMap;
     use tempfile::TempDir;
     use url::Url;
@@ -421,7 +445,8 @@ mod tests {
         let runtime = ApptainerRuntime::new(&root.path().join("runs"), None).unwrap();
         let _ = runtime
             .generate_script(
-                &Default::default(),
+                &ApptainerConfig::default(),
+                None,
                 &ExecuteTaskRequest {
                     id: "example-task",
                     command: "echo hello",
@@ -447,7 +472,6 @@ mod tests {
                     attempt_dir: &root.path().join("0"),
                     temp_dir: &root.path().join("temp"),
                 },
-                empty(),
                 CancellationToken::new(),
             )
             .await
@@ -472,7 +496,8 @@ mod tests {
         let runtime = ApptainerRuntime::new(&root.path().join("runs"), None).unwrap();
         let script = runtime
             .generate_script(
-                &Default::default(),
+                &ApptainerConfig::default(),
+                None,
                 &ExecuteTaskRequest {
                     id: "example-task",
                     command: "echo hello",
@@ -498,7 +523,6 @@ mod tests {
                     attempt_dir: &root.path().join("0"),
                     temp_dir: &root.path().join("temp"),
                 },
-                empty(),
                 CancellationToken::new(),
             )
             .await